third_party/blink/web_tests/inspector-protocol/css/css-set-style-text-expected.txt

The test verifies functionality of protocol method CSS.setStyleTexts.
==== Initial style sheet text ====
#test {
    box-sizing: border-box;
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
}



Running test: testBasicSetStyle
==== Style sheet text ====
#test {
    content: 'EDITED';
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
}


Dumping matched rules: 
*#test* {    regular
    content: 'EDITED'; @[1:4-1:22]
    content: "EDITED"; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[14:8-14:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testSetStyleTwice
==== Style sheet text ====
#test {
    color: green;
    padding: 0 4px;
    cursor: pointer
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
}


Dumping matched rules: 
*#test* {    regular
    color: green; @[1:4-1:17]
    padding: 0 4px; @[2:4-2:19]
    cursor: pointer; @[3:4-4:0]
    color: green; @[undefined-undefined]
    padding-top: 0px; @[undefined-undefined]
    padding-right: 4px; @[undefined-undefined]
    padding-bottom: 0px; @[undefined-undefined]
    padding-left: 4px; @[undefined-undefined]
    cursor: pointer; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[8:4-8:19]
    font-family: "Arial"; @[9:4-9:25]
    color: blue; @[10:4-10:16]
    display: flex; @[11:4-11:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[16:8-16:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 
==== Style sheet text ====
#test {
    color: green;
    padding: 0 6px;
    cursor: pointer
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
}


Dumping matched rules: 
*#test* {    regular
    color: green; @[1:4-1:17]
    padding: 0 6px; @[2:4-2:19]
    cursor: pointer; @[3:4-4:0]
    color: green; @[undefined-undefined]
    padding-top: 0px; @[undefined-undefined]
    padding-right: 6px; @[undefined-undefined]
    padding-bottom: 0px; @[undefined-undefined]
    padding-left: 6px; @[undefined-undefined]
    cursor: pointer; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[8:4-8:19]
    font-family: "Arial"; @[9:4-9:25]
    color: blue; @[10:4-10:16]
    display: flex; @[11:4-11:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[16:8-16:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 
==== Style sheet text ====
#test {
    color: green;
    padding: 0 8px;
    cursor: pointer
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        font-size: 200%;
    }

    #test {
    }
}


Dumping matched rules: 
*#test* {    regular
    color: green; @[1:4-1:17]
    padding: 0 8px; @[2:4-2:19]
    cursor: pointer; @[3:4-4:0]
    color: green; @[undefined-undefined]
    padding-top: 0px; @[undefined-undefined]
    padding-right: 8px; @[undefined-undefined]
    padding-bottom: 0px; @[undefined-undefined]
    padding-left: 8px; @[undefined-undefined]
    cursor: pointer; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[8:4-8:19]
    font-family: "Arial"; @[9:4-9:25]
    color: blue; @[10:4-10:16]
    display: flex; @[11:4-11:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[16:8-16:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testSetStylePoorContent
Expected protocol error: Failed applying edit #0: SyntaxError Style text is not valid.
Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[14:8-14:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testSetStyleOpenBrace
Expected protocol error: Failed applying edit #0: SyntaxError Style text is not valid.
Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[14:8-14:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testSetStyleComment
Expected protocol error: Failed applying edit #0: SyntaxError Style text is not valid.
Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[14:8-14:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testSetStyleInMedia
==== Style sheet text ====
#test {
    box-sizing: border-box;
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {
        content: 'EDITED';
        color: red;
        /** foo */
    }

    #test {
    }
}


Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        content: 'EDITED'; @[14:8-14:26]
        color: red; @[15:8-15:19]
        content: "EDITED"; @[undefined-undefined]
        color: red; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testDeleteStyleBody
==== Style sheet text ====
#test {
    box-sizing: border-box;
}

#test {
    /* resetting some properties */
    line-height: 1;
    font-family: "Arial";
    color: blue;
    display: flex; /* flex FTW! */
}

@media (min-width: 1px) {
    #test {}

    #test {
    }
}


Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testSetStylePoorRange
Expected protocol error: Specified range is out of bounds
Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[14:8-14:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testSetStyleOpenComment
Expected protocol error: Failed applying edit #0: SyntaxError Style text is not valid.
Dumping matched rules: 
*#test* {    regular
    box-sizing: border-box; @[1:4-1:27]
    box-sizing: border-box; @[undefined-undefined]
}
*#test* {    regular
    line-height: 1; @[6:4-6:19]
    font-family: "Arial"; @[7:4-7:25]
    color: blue; @[8:4-8:16]
    display: flex; @[9:4-9:18]
    line-height: 1; @[undefined-undefined]
    font-family: Arial; @[undefined-undefined]
    color: blue; @[undefined-undefined]
    display: flex; @[undefined-undefined]
}
@media (min-width: 1px)
    *#test* {    regular
        font-size: 200%; @[14:8-14:24]
        font-size: 200%; @[undefined-undefined]
    }
@media (min-width: 1px)
    *#test* {    regular
    }
Dumping inherited rules: 

Running test: testSetStyleOfRemovedRule
ERROR: Failed applying edit #0: NotFoundError Source range didn't match existing style source range
